All tests pass. Legacy Serializer object still working
